refactor: update NOT_AVAILABLE cell constant
