+++ src/components/SystemsView/columns/helpers.ts
@@ -1,4 +1,4 @@
-import { NOT_AVAILABLE } from '../../../constants';
+import { NOT_AVAILABLE } from './CellValue';
 
 export const valueOrNotAvailable = <T>(value: T) =>
   value != null && value !== '' ? value : NOT_AVAILABLE;
