+++ src/constants.js
@@ -385,5 +385,4 @@ export const MOVE_SYSTEM_MENU_TEXT = 'Move system';
 export const MOVE_SYSTEMS_MENU_TEXT = 'Move systems';
 export const DEFAULT_DELETE_ERROR_MESSAGE =
   'There was an error processing the request. Please try again.';
-export const NOT_AVAILABLE = 'N/A';
 export const DISABLED_TEXT_COLOR = t_global_text_color_disabled.var;
